+++ client.go
@@ -21,6 +21,7 @@ type Client struct {
 const (
 	endpointGenerate = "generate"
 	endpointEmbed    = "embed"
+	endpointExtract  = "extract"
 	endpointClassify = "classify"
 
 	endpointCheckAPIKey = "check-api-key"
@@ -178,6 +179,23 @@ func (c *Client) Embed(model string, opts EmbedOptions) (*EmbedResponse, error) 
 	}
 
 	ret := &EmbedResponse{}
+	if err := json.Unmarshal(res, ret); err != nil {
+		return nil, err
+	}
+	return ret, nil
+}
+
+// Extracts entities of specified types from the provided text. Each extraction
+// contains a type and a confidence score.
+// See: https://docs.cohere.ai/extract-reference
+// Returns an EmbedResponse object.
+func (c *Client) Extract(model string, opts ExtractOptions) (*ExtractResponse, error) {
+	res, err := c.post(model, endpointExtract, opts)
+	if err != nil {
+		return nil, err
+	}
+
+	ret := &ExtractResponse{}
 	if err := json.Unmarshal(res, ret); err != nil {
 		return nil, err
 	}
