+++ go.mod
@@ -2,9 +2,15 @@ module github.com/cohere-ai/cohere-go
 
 go 1.17
 
-require github.com/cohere-ai/tokenizer v1.0.4
+require (
+	github.com/cohere-ai/tokenizer v1.0.4
+	github.com/stretchr/testify v1.7.1
+)
 
 require (
+	github.com/davecgh/go-spew v1.1.0 // indirect
 	github.com/dlclark/regexp2 v1.4.0 // indirect
 	github.com/pkg/errors v0.9.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
 )
